requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
beautifulsoup4==4.9.3
google==3.0.0
googlesearch-python==1.1.0
playwright==1.27.1
pydantic==1.8.2
python-dotenv==0.19.0
requests==2.25.1
selenium==4.1.0
minify_html==0.15.0
duckduckgo_search==6.3.3
newspaper3k==0.2.8
fake-useragent==1.5.1
webdriver-manager==4.0.2
lxml_html_clean==0.3.1
openai==1.3.0
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
nltk==3.8.1


